+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 ## **WORK IN PROGRESS**
 - feature: checkout any version in complete history 😎
 - feature: show remote versions in github 🥳
+- feature: hugo download tester for all platforms
 - cleanup: refactored sync frontend
 - cleanup: refactored sync api bridge
 - cleanup: refactored sync backend
+- fix: fix hugo downloads after v0.102.x (wrong url's)
 - fix: fix folder sync (synced to wrong path)
 - fix: typo in sync